requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
amqp==5.3.1
asgiref==3.11.0
astroid==4.0.3
attrs==25.4.0
autobahn==25.12.2
Automat==25.4.16
billiard==4.2.4
cbor2==5.8.0
celery==5.6.2
certifi==2026.1.4
cffi==2.0.0
channels==4.3.2
channels_redis==4.3.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cron_descriptor==2.0.6
cryptography==46.0.5
daphne==4.2.1
dill==0.4.1
Django==5.2.11
django-allauth==65.14.0
django-celery-beat==2.8.1
django-stubs==5.2.9
django-stubs-ext==5.2.9
django-timezone-field==7.2.1
django_celery_results==2.6.0
frozenlist==1.8.0
hiredis==3.3.0
http_ece==1.2.1
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
iniconfig==2.3.0
isort==7.0.0
kombu==5.6.2
mccabe==0.7.0
msgpack==1.1.2
multidict==6.7.1
packaging==26.0
pillow==12.1.0
platformdirs==4.5.1
pluggy==1.6.0
prompt_toolkit==3.0.52
propcache==0.4.1
psycopg2-binary==2.9.11
py-ubjson==0.16.1
py-vapid==1.9.4
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
Pygments==2.19.2
pylint==4.0.4
pylint-plugin-utils==0.9.0
pyOpenSSL==25.3.0
pytest==9.0.2
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pywebpush==2.3.0
redis==6.4.0
requests==2.32.5
service-identity==24.2.0
six==1.17.0
sqlparse==0.5.5
tomlkit==0.14.0
Twisted==25.5.0
txaio==25.12.2
types-PyYAML==6.0.12.20250915
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
ujson==5.11.0
urllib3==2.6.3
uv==0.9.28
vine==5.1.0
wcwidth==0.6.0
yarl==1.23.0
zope.interface==8.2
